deleted sync plugin
